get error messages working again

+++ macosx/PrefsController.m
@@ -589,7 +589,7 @@
         return;
     
     NSRect windowRect = [window frame];
-    int difference = [view frame].size.height - [[window contentView] frame].size.height;
+    float difference = [view frame].size.height - [[window contentView] frame].size.height;
     windowRect.origin.y -= difference;
     windowRect.size.height += difference;
